+++ src/util/mod.rs
@@ -1,0 +1,7 @@
+// SPDX-FileCopyrightText: 2026 Manuel Quarneti <mq1@ik.me>
+// SPDX-License-Identifier: GPL-3.0-only
+
+pub mod data_dir;
+pub mod drive_info;
+pub mod http;
+pub mod misc;
